fixed tinterval

+++ src/main/java/com/signomix/common/db/IotDatabaseDao.java
@@ -808,7 +808,7 @@ public class IotDatabaseDao implements IotDatabaseIface {
                 device.setLastSeen(rs.getTimestamp("ts").getTime());
                 device.setState(rs.getDouble("status"));
                 device.setAlertStatus(rs.getInt("alert"));
-                device.setTransmissionInterval(rs.getInt("tinterval"));
+                device.setTransmissionInterval(rs.getLong("tinterval"));
             }
             return device;
         } catch (SQLException e) {
